+++ Demo/Application/Base/PaymentButtonBaseViewController.swift
@@ -7,9 +7,10 @@ class PaymentButtonBaseViewController: BaseViewController {
 
     var heightConstraint: CGFloat?
 
-    private var paymentButton: UIView = UIView()
+    private var paymentButton = UIView()
 
     override init(authorization: String) {
+        // swiftlint:disable:next force_unwrapping
         apiClient = BTAPIClient(authorization: authorization)!
         super.init(authorization: authorization)
     }