+++ Demo/Application/Features/Helpers/CardHelpers.swift
@@ -6,7 +6,7 @@ enum Format {
     case year
 }
 
-class CardHelpers {
+enum CardHelpers {
 
     static func newCard(from cardFormView: BTCardFormView) -> BTCard {
         let card = BTCard()
@@ -34,16 +34,16 @@ class CardHelpers {
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "yy"
 
-        let futureYear = Calendar.current.date(byAdding: .year, value: 3, to: Date())!
+        let futureYear = Calendar.current.date(byAdding: .year, value: 3, to: Date())
 
         switch format {
         case .date:
             let monthString = "12"
-            let yearString = dateFormatter.string(from: futureYear)
+            let yearString = dateFormatter.string(from: futureYear ?? Date.distantFuture)
 
             return "\(monthString)/\(yearString)"
         case .year:
-            return dateFormatter.string(from: futureYear)
+            return dateFormatter.string(from: futureYear ?? Date.distantFuture)
         }
     }
 }