+++ Demo/Application/Features/PayPalMessagingViewController.swift
@@ -14,8 +14,10 @@ class PayPalMessagingViewController: PaymentButtonBaseViewController {
     )
 
     override func viewDidLoad() {
+        super.viewDidLoad()
+        view.removeGestureRecognizer(super.tapToDismissKeyboard)
+
         title = "PayPal Messaging"
-        
         payPalMessagingView.delegate = self
         payPalMessagingView.start(request)
     }
@@ -56,4 +58,3 @@ extension PayPalMessagingViewController: BTPayPalMessagingDelegate {
         progressBlock("DELEGATE: onError fired with \(error.localizedDescription)")
     }
 }
-