+++ CHANGELOG.md
@@ -2,7 +2,22 @@
 
 ## unreleased
 * BraintreePayPal
+  * Add `BTPayPalRequest.userPhoneNumber` optional property
+  
+## 6.24.0 (2024-10-15)
+* BraintreePayPal
+  * Add `BTPayPalRecurringBillingDetails` and `BTPayPalRecurringBillingPlanType` opt-in request objects. Including these details will provide transparency to users on their billing schedule, dates, and amounts, as well as launch a modernized checkout UI.
+
+## 6.23.5 (2024-10-09)
+* BraintreeCore
+  * Fix analytics bug where sessionID value in analytics payload was inaccurate; send separate FPTI POST requests per unique sessionID
+
+## 6.23.4 (2024-09-24)
+* BraintreePayPal
   * Send `isVaultRequest` for App Switch events to PayPal's analytics service (FPTI)
+* BraintreeShopperInsights (BETA)
+  * For analytics, send `experiment` as a parameter to `getRecommendedPaymentMethods` method
+  * For analytics, send `experiment` and `paymentMethodsDisplayed` analytic metrics to FPTI via the button presented methods
 
 ## 6.23.3 (2024-08-12)
 * BraintreeCore